requirements.txt
# =============================================================================
# Image Morph - Python Dependencies
# =============================================================================
# This file contains all Python dependencies for the Image Morph application.
# 
# Installation:
#   pip install -r requirements.txt
#
# For AI upscaling features (RealESRGAN), Python 3.11.x is recommended.
# Python 3.13+ may lack compatible wheels for some AI packages.
# =============================================================================

# -----------------------------------------------------------------------------
# Backend Framework
# -----------------------------------------------------------------------------
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17

# -----------------------------------------------------------------------------
# Database
# -----------------------------------------------------------------------------
sqlalchemy==2.0.36
alembic==1.14.0
aiosqlite==0.20.0

# -----------------------------------------------------------------------------
# Queue Processing (Celery + Redis)
# -----------------------------------------------------------------------------
celery>=5.3.0
redis>=5.0.0

# -----------------------------------------------------------------------------
# Image Processing
# -----------------------------------------------------------------------------
Pillow==11.0.0
numpy>=1.24.0
tqdm

# -----------------------------------------------------------------------------
# Cloud Storage (Optional - Cloudflare R2/S3 compatible)
# -----------------------------------------------------------------------------
boto3==1.35.0
botocore==1.35.0

# -----------------------------------------------------------------------------
# Configuration & Validation
# -----------------------------------------------------------------------------
pydantic==2.9.0
pydantic-settings==2.6.0
python-dotenv==1.0.1
aiofiles==24.1.0

# -----------------------------------------------------------------------------
# Security
# -----------------------------------------------------------------------------
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# =============================================================================
# OPTIONAL: AI Upscaling Dependencies (Real-ESRGAN Stack)
# =============================================================================
# Uncomment the following lines to enable AI-powered image upscaling.
# These packages require Python < 3.13 and have complex dependencies.
#
# Installation order matters:
#   1. Install torch/torchvision first (CPU or CUDA version)
#   2. Install basicsr from git: pip install git+https://github.com/xinntao/BasicSR.git@master
#   3. Then install the packages below
#
# For automated installation, use: python install_deps.py
# =============================================================================

# realesrgan>=0.3.0; python_version < "3.13"
# facexlib>=0.3.0; python_version < "3.13"
# gfpgan>=1.3.8; python_version < "3.13"
# filterpy; python_version < "3.13"
